Add live template processing.

+++ md.py
@@ -6,6 +6,9 @@
 
 import os
 from os.path import join
+from os import walk
+import re
+
 import time
 import logging
 
@@ -41,13 +44,18 @@ def process_file(file):
 
 class Handler(FileSystemEventHandler):
     def on_any_event(self, event):
-        if type(event) is FileModifiedEvent or type(event) is FileCreatedEvent:
+        if (type(event) is FileModifiedEvent or type(event) is FileCreatedEvent) and event.src_path.startswith("md" + os.path.sep):
             logging.info("File created or changed! Path: {path}".format(path=event.src_path))
             process_file(event.src_path)
-        elif type(event) is FileDeletedEvent:
+        elif (type(event) is FileDeletedEvent) and event.src_path.startswith("md" + os.path.sep):
             logging.info("File deleted! Path: {path}".format(path=event.src_path))
             if os.path.exists(event.src_path):
                 os.remove(event.src_path.replace("/md/", "/pages/"))
+        elif (type(event) is FileModifiedEvent) and len(event.src_path) >= len("template.html") and event.src_path[len(event.src_path)-len("template.html"):] == "template.html":
+            logging.info("Template changed, reprocessing all files.")
+            for (dirpath, dirnames, filenames) in walk(mypath):
+                for x in filenames:
+                    process_file(dirpath + os.path.sep + x)
 
 
 if __name__ == "__main__":
@@ -55,7 +63,7 @@ if __name__ == "__main__":
                         format='%(asctime)s - %(message)s',
                         datefmt='%Y-%m-%d %H:%M:%S')
     observer = Observer()
-    observer.schedule(Handler(), mypath, recursive=True)
+    observer.schedule(Handler(), ".", recursive=True)
     observer.start()
     try:
         while True:
